Database.Record.TH

Description

This module defines templates for Haskell record type and type class instances to map between list of untyped SQL type and Haskell record type.

Generate all templates about record

defineRecord

Arguments

:: TypeQ
SQL value type

-> ConName
Record type name

-> [(VarName, TypeQ)]
Column schema

-> [Name]
Record derivings

-> Q [Dec]
Result declarations

All templates for record type.

defineRecordWithConfig

Arguments

:: TypeQ
SQL value type

-> NameConfig
name rule config

-> String
Schema name

-> String
Table name

-> [(String, TypeQ)]
Column names and types

-> [Name]
Record derivings

-> Q [Dec]
Result declarations

All templates for record type with configured names.

Table constraint specified by key

defineHasColumnConstraintInstance

Arguments

:: TypeQ
Type which represent constraint type

-> TypeQ
Type constructor of record

-> Int
Key index which specifies this constraint

-> Q [Dec]
Result declaration template

Template of HasColumnConstraint instance.

defineHasPrimaryConstraintInstanceDerived

Arguments

:: TypeQ
Type constructor of record

-> Q [Dec]
Result declaration template

Template of HasKeyConstraint instance.

defineHasPrimaryKeyInstance

Arguments

:: TypeQ
Type constructor of record

-> [Int]
Key index which specifies this constraint

-> Q [Dec]
Declaration of primary key constraint instance

Template of HasColumnConstraint Primary instance.

defineHasNotNullKeyInstance

Arguments

:: TypeQ
Type constructor of record

-> Int
Key index which specifies this constraint

-> Q [Dec]
Declaration of not null key constraint instance

Template of HasColumnConstraint NotNull instance.

Record type

defineRecordType

Arguments

:: ConName
Name of the data type of table record type.

-> [(VarName, TypeQ)]
List of columns in the table. Must be legal, properly cased record columns.

-> [Name]
Deriving type class names.

-> Q [Dec]
The data type record declaration.

Record type declaration template.

defineRecordTypeWithConfig :: NameConfig -> String -> String -> [(String, TypeQ)] -> [Name] -> Q [Dec]

Record type declaration template with configured names.

Function declarations against defined record types

defineColumnOffsets

Arguments

:: ConName
Record type constructor.

-> [TypeQ]
Types of record columns.

-> Q [Dec]
Declaration of PersistableWidth instance.

Column offset array and PersistableWidth instance declaration.

recordWidthTemplate

Arguments

:: TypeQ
Record type constructor.

-> ExpQ
Expression to get record width.

Record type width expression template.

Reify

reifyRecordType :: Name -> Q ((TypeQ, ExpQ), (Maybe [Name], [TypeQ]))

Low-level reify interface for record type name.

Templates about record name

data NameConfig

NameConfig type to customize names of expanded record templates.

defaultNameConfig :: NameConfig

Default implementation of NameConfig type. To customize this, use record update syntax.

recordTypeName :: NameConfig -> String -> String -> ConName

Make record type symbol name from schema name and table name in SQL

columnName :: NameConfig -> String -> String -> VarName

Make column variable symbol name from table name and column name in SQL

recordTemplate

Arguments

:: NameConfig
name rule config

-> String
Schema name string in SQL

-> String
Table name string in SQL

-> (TypeQ, ExpQ)
Record type and data constructor

Record constructor templates from SQL table name String.

columnOffsetsVarNameDefault

Arguments

:: Name
Table type name

-> VarName
Result expression variable name

Variable expression of record column offset array.

Not nullable single column type

deriveNotNullType :: TypeQ -> Q [Dec]

Templates for single column value type.

Template for tuple types

defineTupleInstances :: Int -> Q [Dec]

Template to define tuple instances of persistable-record classes.